dashboard/requirements.txt
fastapi>=0.115
uvicorn>=0.40
asyncpg>=0.30
python-dotenv>=1.0
jinja2>=3.1
sentence-transformers>=5.0
